Fix ever-increasing Postgres AWS RDS storage issue (#1)
The [Postgres WAL (Write-Ahead Log)][1] was growing permanently after
starting Debezium's CDC (Change Data Capture) on Confluent Cloud's
Kafka. The reason was that our AWS RDS database is mostly idle, but
under the hood AWS RDS writes to its own system tables every 5 minutes
(see the [tip in Debezium's documentation][2] for more info).  This was
causing the WAL to keep growing as Debezium was not being invoked and
therefore the database was not reconciling the WAL.

The fix in this commit is to add a heartbeat table to the database
and to configure Confluent Cloud / Kakfa / Debezium to update this table
via the heartbeat, while also configuring the update to be emitted via
Debezium. That is to say we include the heartbeat table in the list of
tables that Debezium connects to (ensuring that the heartbeat events are
emitted and therefore that the database reconciles the WAL, recapturing
the disk space).

You can verify that this fix is working properly by running the SQL
query below against your AWS database and then running it again after,
say, 30 minutes.  You will see that the `replicationslotlag` is much
the same both times (in testing it has appeared no higher than 192MB,
which is fine), instead of being significantly larger the second
time.

`SELECT slot_name, pg_size_pretty(pg_wal_lsn_diff(pg_current_wal_lsn(),restart_lsn)) AS replicationSlotLag,
active FROM pg_replication_slots ;`

[1]: https://www.postgresql.org/docs/current/wal-intro.html
[2]: https://debezium.io/documentation/reference/1.0/connectors/postgresql.html#wal-disk-space